refactor(auth): move password-related functions to a new utils module for better organization The password hashing and comparison functions are moved from the auth module to a new utils module. This change improves code organization and separation of concerns, making it easier to manage authentication-related utilities in one place. Additionally, it enhances code readability and maintainability.
