+++ prisma/production-seed/users.ts
@@ -1,5 +1,5 @@
 import { prisma } from "../seed/prisma-client";
-import { hashPassword } from "../../src/global/auth/auth";
+import { hashPassword } from "../../src/global/auth/utils";
 import { getRoleId } from "../seed/users";
 import { generateGravatarUrl } from "../seed/utils";
