+++ prisma/seed/users.ts
@@ -1,4 +1,4 @@
-import { hashPassword } from "../../src/global/auth/auth";
+import { hashPassword } from "../../src/global/auth/utils";
 import { prisma } from "./prisma-client";
 import { generateGravatarUrl } from "./utils";
