+++ src/auth/discord-auth.service.ts
@@ -6,7 +6,7 @@ import {
 import { IAuthProvider } from "../global/interfaces/oauth.interface";
 import { PrismaService } from "../prisma/prisma.service";
 import { DiscordUser } from "../global/types/auth.types";
-import { generatePasswordHash } from "../global/auth/auth";
+import { generatePasswordHash } from "../global/auth/utils";
 
 import { OAuthConfig } from "@/config/Oauth/oauthConfig.interface";
 @Injectable()
