+++ test/auth.e2e-spec.ts
@@ -11,7 +11,7 @@ import {
     loginAndGetTokens,
 } from "./utils";
 import { PrismaService } from "../src/prisma/prisma.service";
-import { comparePassword } from "../src/global/auth/auth";
+import { comparePassword } from "../src/global/auth/utils";
 import { CASLForbiddenExceptionFilter } from "../src/exception-filters/casl-forbidden-exception.filter";
 
 import { AuthConfig } from "../src/config/auth/auth.interface";
